+++ jgclark.QuickCapture/src/quickCapture.js
@@ -302,7 +302,7 @@ export async function addTextToNoteHeading(
 ): Promise<void> {
   try {
     logDebug(pluginJson, `starting /qalh with arg0 '${noteTitleArg}' arg1 '${headingArg}' arg2 ${textArg != null ? '<text defined>' : '<text undefined>'} arg3 ${headingLevelArg}`)
-    const config = (await getQuickCaptureSettings())||{}
+    const config = await getQuickCaptureSettings()
 
     // Start a longish sort job in the background
     CommandBar.onAsyncThread()